requirements-dev.txt
# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit>=3.5.0

# Core dependencies needed for development scripts
pydantic>=2.9
structlog==24.1.0
httpx>=0.27
GitPython>=3.1.40  # For cloning repositories
openai>=1.12.0  # For collecting OpenAI responses